이미지 추가, 로딩 화면 이미지 문구 추가, 픽셀 수정, 앱 이름 수정, 오류 수정

+++ flutterproject/lib/screens/recommend_cloth.dart
@@ -26,7 +26,7 @@ class _RecommendClothState extends State<RecommendCloth> {
   }
 
   void updateData(dynamic weatherData) {
-    double temp2 = weatherData['main']['temp'];
+    double temp2 = weatherData['main']['temp'].toDouble();
     temp = temp2.toInt();
     cityName = weatherData['name'];
     int condition = weatherData['weather'][0]['id'];
